started tests for post request to login
