+++ package.json
@@ -18,6 +18,7 @@
     "helmet": "^5.0.2",
     "jsonwebtoken": "^8.5.1",
     "mongodb-memory-server": "^8.3.0",
+    "mongodb-memory-server-core": "^8.3.0",
     "mongoose": "^6.2.1",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.15",
